refactor: add Fragment for inject key to dateCell
